Move ch-showcase to "src/showcase/assets/components/" folder
