+++ src/showcase/assets/components/readme.md
